python/requirements.txt
requests==2.26.0
PyYAML==6.0
Flask==2.0.2